ecommerce/user_service/requirements.txt
Django==5.0.4
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
mongoengine==0.27.0
pymongo==4.6.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
bcrypt==4.0.1
psycopg2-binary==2.9.9
Pillow>=10.0